requirements.txt
Django>=1.8,<1.10
Celery==3.1.20
django-celery==3.1.17
Scrapy>=1.1,<1.2
scrapy-djangoitem>=1.1.1,<1.2
scrapyd>=1.1,<1.2
jsonpath-rw>=1.4
django-dynamic-scraper
